+++ src/components/store/meal-recruit/MealRecruitCard.tsx
@@ -1,5 +1,7 @@
 import { useCallback } from 'react';
 
+import { convertUtcToLocal } from '@/utils/convertUtcToLocal';
+
 import { useDialogContext } from '@/hooks';
 import { MealPosts } from '@/types/store/storeMealPostDto';
 import { formatDate } from '@/utils';
@@ -26,21 +28,6 @@ export default function MealRecruitCard({
 
   const isDisabled = false;
 
-  // TODO - 공통 유틸함수로 분리 (서버에서 받은 UTC 시간 문자열을 로컬 시간으로 변환하는 로직)
-  // UTC 시간 문자열
-  const utcTimeString = slideItem.appointmentTime;
-
-  // UTC 시간을 구성하는 연, 월, 일, 시, 분, 초 값을 분리
-  const [year, month, day, hour, minute, second] = utcTimeString
-    .split(/[-T:]/)
-    .map(Number);
-
-  // Date.UTC를 사용하여 UTC 시간을 밀리초(Timestamp)로 변환
-  const utcTimestamp = Date.UTC(year, month - 1, day, hour, minute, second);
-
-  // 변환된 Timestamp 값으로 Date 객체 생성 (로컬 시간으로 변환됨)
-  const localDate = new Date(utcTimestamp);
-
   return (
     <div className="h-full w-full text-start">
       <div
@@ -61,7 +48,7 @@ export default function MealRecruitCard({
             </span>
             <span className={`${isDisabled && 'text-gray1'}`}>
               {slideItem
-                ? `${formatDate(localDate, 'yyyy.MM.dd a h시 mm분')}`
+                ? `${formatDate(convertUtcToLocal(slideItem.appointmentTime), 'yyyy.MM.dd a h시 mm분')}`
                 : '-'}
             </span>
           </p>
